// src/utils/types.ts
export type DemoItem = {
  id: number; // Unique identifier for the demo item
  title: string; // Title of the demo item
  description: string; // Description of the demo item
  imageUrl: string; // URL of the image associated with the demo item
  link: string;
  project: string; // project label   }
};